% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ktest.R
\name{Ktest}
\alias{Ktest}
\title{Hypothesis test for the Kurtosis  of Independent component models}
\usage{
Ktest(X, k0, alternative = c("two.sided", "less", "greater"))
}
\arguments{
\item{X}{data matrix of dimension n*p}

\item{k0}{the value of kurtosis under the null hypothesis}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of '"two.sided"' (default), '"greater"' or '"less"'.}
}
\value{
A list with class \item{htest}
containing the following components:
\item{statistic}{the list containing kurtosis and its standard deviation and the test statistic.}
\item{p.value}{the p-value for the test}
\item{alternative}{a character string describing the alternative hypothesis}
\item{data.name}{name of the data argument}
}
\description{
Hypothesis test for the Kurtosis  of Independent component models
}
