% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurtICM.R
\name{kurtICM}
\alias{kurtICM}
\title{Efficient algorithm for Kurtosis estimation of Independent component model}
\usage{
kurtICM(X)
}
\arguments{
\item{X}{data matrix of dimension n*p}
}
\value{
A list with components
\item{ke}{The esitimation value of kurtosis}
\item{sde}{The estimation value of standard deviation of ka}
}
\description{
Efficient algorithm for Kurtosis estimation of Independent component model
}
