% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KurICM_E.R
\name{kurtosisU}
\alias{kurtosisU}
\title{Efficient algorithm for Kurtosis estimation of Independent component model}
\usage{
kurtosisU(X)
}
\arguments{
\item{X}{data matrix of dimension n*p}
}
\value{
A list with components
\item{ka}{The esitimation value of kurtosis}
\item{tau}{The estimation value of tau in the reference.}
}
\description{
Efficient algorithm for Kurtosis estimation of Independent component model
}
